bissel Conetnt Browser verbessert man kann scahen importioeren

+++ runEditor.ps1
@@ -1,4 +1,11 @@
-Remove-Item -Path "ScriptingTest/build/libs/ScriptingTest.jar" -ErrorAction SilentlyContinue
+param (
+    [switch]$noRecompileProject
+)
 
-.\gradlew.bat :ScriptingTest:jar
-.\gradlew.bat :Editor:runImGuiTest
+if ($noRecompileProject) {
+    .\gradlew.bat :Editor:runEditor
+} else {
+    Remove-Item -Path "ScriptingTest/build/libs/ScriptingTest.jar" -ErrorAction SilentlyContinue
+    .\gradlew.bat :ScriptingTest:jar
+    .\gradlew.bat :Editor:runEditor
+}
